migrate to next/mdx for typechecking
