+++ apps/docs/app/docs/Installation/content.mdx
@@ -4,7 +4,7 @@ Diffs is
 [published as an npm package](https://www.npmjs.com/package/@pierre/diffs).
 Install Diffs with the package manager of your choice:
 
-<PackageManagerTabs installationExamples={installationExamples} />
+<PackageManagerTabs installationExamples={props.installationExamples} />
 
 ### Package Exports [toc-ignore]
