+++ apps/docs/app/docs/SSR/content.mdx
@@ -22,11 +22,11 @@ corresponding React component for automatic hydration.
 
 #### Server Component
 
-<DocsCodeExample {...usageServer} />
+<DocsCodeExample {...props.usageServer} />
 
 #### Client Component
 
-<DocsCodeExample {...usageClient} />
+<DocsCodeExample {...props.usageClient} />
 
 ### Preloaders
 
@@ -39,7 +39,7 @@ Preloads a single file with syntax highlighting (no diff). Use this when you
 want to render a file without any diff context. Spread into the `File`
 component.
 
-<DocsCodeExample {...preloadFileResult} />
+<DocsCodeExample {...props.preloadFileResult} />
 
 #### preloadFileDiff
 
@@ -47,7 +47,7 @@ Preloads a diff from a `FileDiffMetadata` object. Use this when you already have
 parsed diff metadata (e.g., from `parseDiffFromFile` or `parsePatchFiles`).
 Spread into the `FileDiff` component.
 
-<DocsCodeExample {...preloadFileDiff} />
+<DocsCodeExample {...props.preloadFileDiff} />
 
 #### preloadMultiFileDiff
 
@@ -55,14 +55,14 @@ Preloads a diff directly from old and new file contents. This is the simplest
 option when you have the raw file contents and want to generate a diff. Spread
 into the `MultiFileDiff` component.
 
-<DocsCodeExample {...preloadMultiFileDiff} />
+<DocsCodeExample {...props.preloadMultiFileDiff} />
 
 #### preloadPatchDiff
 
 Preloads a diff from a unified patch string for a single file. Use this when you
 have a patch in unified diff format. Spread into the `PatchDiff` component.
 
-<DocsCodeExample {...preloadPatchDiff} />
+<DocsCodeExample {...props.preloadPatchDiff} />
 
 #### preloadPatchFile
 
@@ -70,4 +70,4 @@ Preloads multiple diffs from a multi-file patch string. Returns an array of
 results, one for each file in the patch. Each result can be spread into a
 `FileDiff` component.
 
-<DocsCodeExample {...preloadPatchFile} />
+<DocsCodeExample {...props.preloadPatchFile} />
