+++ apps/docs/app/shadcn/page.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { File } from '@pierre/diffs/react';
 
 import { Examples } from './Examples';
