+++ apps/docs/components/MDXClientComponents.tsx
@@ -1,0 +1,5 @@
+'use client';
+
+// Re-export client components from @pierre/diffs/react for use in MDX
+// This file is marked 'use client' so Next.js knows these are client components
+export { MultiFileDiff } from '@pierre/diffs/react';
