+++ apps/docs/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "build": "bun run build:deps && bun run build:next",
-    "build:next": "next build --turbopack",
+    "build:next": "next build --webpack",
     "build:deps": "bun run build:deps:diffs && bun run build:deps:elements-next && bun run build:deps:registry",
     "build:deps:diffs": "output=$(cd ../../packages/diffs && bun run build 2>&1) && echo '[diffs] Successfully cleaned and built.' || (echo \"$output\" >&2 && exit 1)",
     "build:deps:elements-next": "(cd ../../packages/storage-elements-next && bun run build)",
@@ -43,7 +43,9 @@
     "hast-util-to-string": "catalog:",
     "lucide-react": "catalog:",
     "next": "catalog:",
-    "next-mdx-remote": "catalog:",
+    "@mdx-js/mdx": "catalog:",
+    "@mdx-js/react": "catalog:",
+    "@next/mdx": "catalog:",
     "next-themes": "catalog:",
     "postcss": "catalog:",
     "postcss-import": "catalog:",
