+++ package.json
@@ -47,7 +47,10 @@
       "@eslint/eslintrc": "3.3.1",
       "@eslint/js": "9.36.0",
       "@icons-pack/react-simple-icons": "13.8.0",
+      "@mdx-js/mdx": "3.1.0",
+      "@mdx-js/react": "3.1.0",
       "@next/eslint-plugin-next": "16.0.7",
+      "@next/mdx": "16.0.7",
       "@octokit/app": "16.1.1",
       "@octokit/auth-app": "8.1.1",
       "@octokit/openapi-types": "26.0.0",
